+++ starlite/plugins/sql_alchemy.py
@@ -2,7 +2,7 @@ from datetime import date, datetime, time, timedelta
 from decimal import Decimal
 from inspect import isclass
 from ipaddress import IPv4Network, IPv6Network
-from typing import Any, Callable, Dict, List, Tuple, Union
+from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 from uuid import UUID
 
 from pydantic import BaseModel, Json, conint, constr, create_model
@@ -16,7 +16,7 @@ from starlite.exceptions import (
 from starlite.plugins.base import PluginProtocol
 
 try:
-    from sqlalchemy import Table, inspect
+    from sqlalchemy import inspect
     from sqlalchemy import types as sqlalchemy_type
     from sqlalchemy.dialects import (
         firebird,
@@ -33,7 +33,7 @@ except ImportError as exc:  # pragma: no cover
     raise MissingDependencyException("sqlalchemy is not installed") from exc
 
 
-class SQLAlchemyPlugin(PluginProtocol[Union[DeclarativeMeta, Table]]):
+class SQLAlchemyPlugin(PluginProtocol[DeclarativeMeta]):
     def __init__(self) -> None:
         # a map object that maps SQLAlchemy entity qualnames to pydantic BaseModel subclasses
         self.model_namespace_map: Dict[str, Type[BaseModel]] = {}
@@ -284,15 +284,19 @@ class SQLAlchemyPlugin(PluginProtocol[Union[DeclarativeMeta, Table]]):
                 # to avoid duplication of pydantic models, we are using forward refs
                 # see: https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
                 for name, relationship_property in mapper.relationships.items():
-                    related_entity_class = relationship_property.mapper.class_
-                    related_model_name = related_entity_class.__qualname__
-                    if relationship_property.uselist:
-                        field_definitions[name] = (List[related_model_name], ...)  # type: ignore
+                    if not relationship_property.back_populates and not relationship_property.backref:
+                        related_entity_class = relationship_property.mapper.class_
+                        related_model_name = related_entity_class.__qualname__
+                        if relationship_property.uselist:
+                            field_definitions[name] = (Optional[List[related_model_name]], None)  # type: ignore
+                        else:
+                            field_definitions[name] = (Optional[related_model_name], None)
+                        # if the names are not identical, these are different SQLAlchemy entities
+                        if related_model_name != model_name and related_model_name not in self.model_namespace_map:
+                            related_entity_classes.append(related_entity_class)
+                    # we are treating back-references as any to avoid infinite recursion
                     else:
-                        field_definitions[name] = (related_model_name, ...)
-                    # if the names are not identical, these are different SQLAlchemy entities
-                    if related_model_name != model_name and related_model_name not in self.model_namespace_map:
-                        related_entity_classes.append(related_entity_class)
+                        field_definitions[name] = (Any, None)
             self.model_namespace_map[model_name] = create_model(model_name, **field_definitions)
             for related_entity_class in related_entity_classes:
                 self.to_pydantic_model_class(model_class=related_entity_class)
@@ -300,7 +304,9 @@ class SQLAlchemyPlugin(PluginProtocol[Union[DeclarativeMeta, Table]]):
         model.update_forward_refs(**self.model_namespace_map)
         return model
 
-    def from_pydantic_model_instance(self, model_class: DeclarativeMeta, pydantic_model_instance: BaseModel) -> Any:
+    def from_pydantic_model_instance(
+        self, model_class: Type[DeclarativeMeta], pydantic_model_instance: BaseModel
+    ) -> Any:
         """
         Create an instance of a given model_class using the values stored in the given pydantic_model_instance
         """
@@ -318,3 +324,9 @@ class SQLAlchemyPlugin(PluginProtocol[Union[DeclarativeMeta, Table]]):
         for field in pydantic_model.__fields__:
             kwargs[field] = getattr(model_instance, field)
         return pydantic_model(**kwargs).dict()
+
+    def from_dict(self, model_class: Type[DeclarativeMeta], **kwargs: Any) -> DeclarativeMeta:
+        """
+        Given a dictionary of kwargs, return an instance of the given model_class
+        """
+        return model_class(**kwargs)